demo-aws-master/django-app/requirements.txt
asgiref==3.8.1
boto3==1.37.4
botocore==1.37.4
Django==5.1.6
django-cors-headers==4.7.0
djangorestframework==3.15.2
jmespath==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.11.3
six==1.17.0
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.3.0
